Fixed float-int errors on min/max function parameters
git-svn-id: https://vtr-verilog-to-routing.googlecode.com/svn/trunk@1229 8e3573b8-cf2c-4f14-ef6d-137439e28b8b

+++ vpr/SRC/route/route_breadth_first.c
@@ -81,7 +81,7 @@ boolean try_breadth_first_route(struct s_router_opts router_opts,
 		else
 			pres_fac *= router_opts.pres_fac_mult;
 
-		pres_fac = min(pres_fac, HUGE_POSITIVE_FLOAT / 1e5);
+		pres_fac = min(pres_fac, static_cast<float>(HUGE_POSITIVE_FLOAT / 1e5));
 
 		pathfinder_update_cost(pres_fac, router_opts.acc_fac);
 	}
